devices/requirements.txt
cryptography==3.2
google-api-python-client==1.9.3
google-auth-httplib2==0.0.3
google-auth==1.18.0
google-cloud-pubsub==1.6.0
google-cloud-iot==1.0.0
grpc-google-iam-v1==0.12.3
pyjwt==1.7.1
paho-mqtt==1.5.0
obd
opencv-contrib-python
picamera
ELM327-emulator